refactor: move global volume controller icon to original component
